2019: day 9, fixed intcode lib and refactored earlier days

+++ src/2019/day02/test_day02.cpp
@@ -3,31 +3,31 @@
 #include AOC_HEADER
 
 TEST(AOC_TEST, test_1) {
-    std::vector<int> inp = {1, 9, 10, 3, 2, 3, 11, 0, 99, 30, 40, 50};
+    std::vector<int64_t> inp = {1, 9, 10, 3, 2, 3, 11, 0, 99, 30, 40, 50};
     auto out = solve_1(inp);
     ASSERT_EQ(out, 3500);
 }
 
 TEST(AOC_TEST, test_2) {
-    std::vector<int> inp = {1, 0, 0, 0, 99};
+    std::vector<int64_t> inp = {1, 0, 0, 0, 99};
     auto out = solve_1(inp);
     ASSERT_EQ(out, 2);
 }
 
 TEST(AOC_TEST, test_3) {
-    std::vector<int> inp = {2, 3, 0, 3, 99};
+    std::vector<int64_t> inp = {2, 3, 0, 3, 99};
     auto out = solve_1(inp);
     ASSERT_EQ(out, 2);
 }
 
 TEST(AOC_TEST, test_4) {
-    std::vector<int> inp = {2, 4, 4, 5, 99, 0};
+    std::vector<int64_t> inp = {2, 4, 4, 5, 99, 0};
     auto out = solve_1(inp);
     ASSERT_EQ(out, 2);
 }
 
 TEST(AOC_TEST, test_5) {
-    std::vector<int> inp = {1, 1, 1, 4, 99, 5, 6, 0, 99};
+    std::vector<int64_t> inp = {1, 1, 1, 4, 99, 5, 6, 0, 99};
     auto out = solve_1(inp);
     ASSERT_EQ(out, 30);
 }
